requirements.txt
phidata 
python-dotenv
streamlit
google-generativeai
duckduckgo-search
